BL-72 <updated all files for cleaner merge>
