+++ src/components/pages/Table/RenderTableContainer.jsx
@@ -5,7 +5,7 @@ import { columns } from '../../../data/columns';
 import { makeStyles } from '@material-ui/core';
 import { connect } from 'react-redux';
 import { getAllData } from '../../../state/actions';
-import SearchBar from './SearchBar/SearchBar';
+import SearchBar from './SearchBar';
 
 const useStyles = makeStyles({
   container: {
@@ -15,7 +15,7 @@ const useStyles = makeStyles({
 
 //data will come in as props and not be hard coded
 function RenderTablePage(props) {
-  const { asylum, getAllData, filteredCount, filteredData } = props;
+  const { cases, getAllData, filteredCount, filteredData } = props;
 
   const classes = useStyles();
 
@@ -30,9 +30,9 @@ function RenderTablePage(props) {
       <Table
         className={classes.container}
         bordered={true}
-        loading={asylum.length === 0 ? true : false}
+        loading={cases.length === 0 ? true : false}
         columns={columns}
-        dataSource={filteredCount === 0 ? asylum : filteredData}
+        dataSource={filteredCount === 0 ? cases : filteredData}
         scroll={{ y: 550 }}
         pagination={{
           position: ['bottomCenter'],
@@ -45,21 +45,14 @@ function RenderTablePage(props) {
 }
 
 const mapStateToProps = state => {
-  const filteredReducer = state.filterReducer;
+  const reducerState = state.dataReducer;
+  const filteredReducer = reducerState.filteredReducer;
 
   return {
-    asylum: state.dataReducer,
+    cases: reducerState.casesReducer,
     filteredData: filteredReducer.data,
     filteredCount: filteredReducer.count,
   };
 };
-
-// const mapDispatchToProps = dispatch => {
-//   return {
-//     getAllData: () => {
-//       dispatch(getAllData(data));
-//     },
-//   };
-// };
 
 export default connect(mapStateToProps, { getAllData })(RenderTablePage);
