+++ src/state/reducers/dataReducer.js
@@ -1,8 +1,23 @@
-import { GET_DATA } from '../constants';
+import { combineReducers } from 'redux';
+import {
+  GET_DATA,
+  FILTER_SEARCH,
+  RESET_DATA,
+  SHOW_ADVANCED_SEARCH,
+  ADVANCED_SEARCH,
+} from '../constants';
 
-const initialAsylumData = [];
+// States
+const initialCasesData = [];
 
-function asylumReducer(state = initialAsylumData, action) {
+const initialFilteredStates = {
+  count: 0,
+  data: [],
+  hideAdvanced: true,
+};
+
+// Reducers
+function casesReducer(state = initialCasesData, action) {
   switch (action.type) {
     case GET_DATA: {
       return action.payload;
@@ -13,4 +28,24 @@ function asylumReducer(state = initialAsylumData, action) {
   }
 }
 
-export default asylumReducer;
+function filteredReducer(state = initialFilteredStates, action) {
+  switch (action.type) {
+    case FILTER_SEARCH: {
+      return { ...state, data: action.payload, count: state.count + 1 };
+    }
+    case ADVANCED_SEARCH: {
+      return { ...state, data: action.payload, count: state.count + 1 };
+    }
+    case RESET_DATA: {
+      return initialFilteredStates;
+    }
+    case SHOW_ADVANCED_SEARCH: {
+      return { ...state, hideAdvanced: action.payload };
+    }
+    default: {
+      return state;
+    }
+  }
+}
+
+export default combineReducers({ casesReducer, filteredReducer });
