+++ src/state/reducers/index.js
@@ -2,6 +2,5 @@
 // This allows for the simplification of flow when importing reducers into your actions throughout your app.
 import { combineReducers } from 'redux';
 import dataReducer from './dataReducer';
-import filterReducer from './filterReducer';
 
-export default combineReducers({ dataReducer, filterReducer });
+export default combineReducers({ dataReducer });
